Move lid logic to its own file
