+++ src/daemon.rs
@@ -14,7 +14,8 @@ use super::graph::{Graph, Grapher};
 use super::logger;
 use super::logger::Interface;
 use super::network::{hook, listen};
-use super::power::{read_lid_state, read_power_source, LidState};
+use super::power::lid::{read_lid_state, LidState};
+use super::power::read_power_source;
 use super::settings::{GraphType, Settings};
 use super::system::{
     check_available_governors, check_cpu_freq, check_cpu_temperature, check_cpu_usage,
